sli_stage_A4_GH/smart-li_2-0_dev/requirements.txt
Flask==2.2.3
Flask_WTF==1.1.1
openpyxl==3.1.2
pandas==2.0.0
psycopg2==2.9.6
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==2.3.3
xlsxwriter==3.1.0